"modifyClassStudents handles checks for differences between columns and database.
It checks if classes are listed in either of the columns, if there are,
it then looks for differences between what is currently in the database
and what is now showing in each of the columns based on the user input.

Depending on differences, it will add classes which were not previously
listed for that student, or remove classes that were previously listed
but have been removed.